requirements.txt
asgiref==3.5.1
astroid==2.11.5
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.4
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
PyJWT==2.4.0
pylint==2.13.8
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.9.5
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
wrapt==1.14.1
